requirements.txt
pyyaml>=6.0
python-docx>=1.1.0
pymorphy3>=2.0
pymorphy3-dicts-ru>=2.0
vobject>=0.9.7
pytest>=8.0
pytest-cov>=5.0
ruff>=0.4
vulture>=2.11
mypy>=1.10
